# Remove non-lazy If from Filter implementation