--- a/tests/hspec/SetSpec.hs
+++ b/tests/hspec/SetSpec.hs
@@ -1,6 +1,7 @@
 {-# LANGUAGE TypeApplications #-}
 {-# LANGUAGE DataKinds #-}
 {-# LANGUAGE TypeOperators #-}
+{-# OPTIONS_GHC -fno-pre-inlining #-}
 
 module SetSpec where
 
@@ -17,3 +18,5 @@ spec = do
       barfooStr `shouldBe` "str1"
     it "Assert non-membership of a type not in a set at runtime" $ do
       barHasNat1 `shouldBe` False
+    it "Union of large sets should run in reasonable time" $ do
+      (r0_9 `union` r10_19) `shouldBe` (r0_9 `append` r10_19)