Addresses #17. Moving the `Filter` branching to top level type family patterns does appear to recover linear runtime. A rudimentary benchmark compiling the `hitmen` executable from https://github.com/jmorag/databass on my machine.
```
# current main
cabal run hitmen > b340c822bc25e75d81114327dd9a9161943d9111-out.txt  111.82s user 5.57s system 96% cpu 2:01.38 total
# with the new filter
cabal run hitmen > 1617525181eb4214ea9b65343af4351e5c07d9f5-out.txt  22.38s user 1.09s system 93% cpu 25.175 total
```