@jmorag 2022-01-23
FWIW, I'm testing compilation times with `-O0`. Anything higher still takes forever.