@jmorag 2022-01-23
With the manifesto executable from the same repo `cabal clean; time cabal build manifesto`

|    | New filter | Old filter |
| - | ---------- | -------- | 
| O0 | 10.74s user 0.96s system 88% cpu 13.209 total | 11.33s user 0.98s system 89% cpu 13.734 total |
| O1 | 21.16s user 1.06s system 93% cpu 23.662 total | 85.62s user 3.39s system 96% cpu 1:31.96 total |