@jmorag 2022-01-23
These are not the most scientific benchmarks but they almost don't have to be with that dramatic a difference. I'm trying to compile the 9x9 unification example from the issue on O1 but it's not terminating any time soon. Maybe if I leave it overnight...