@jmorag 2022-01-24
> Maybe if I leave it overnight...

Nope, RAM ran out before GHC finished. Investigated more and the culprit seems to be `-fpre-inlining`. Disabling it in the module where you call `union` on large sets brings compile time down to something reasonable. Added a test case to `SetSpec` to this effect.